Split NSX-V and NSX-T tests to use 'nsxt' tag
